server/requirements.txt
accelerate>=0.26.0
datasets>=2.0.0 
evaluate >= 0.4.0
flask>=2.0.1
flask-cors>=4.0.0
openpyxl>=3.1.2 
selenium==4.18.1
spacy>=3.0.0
scikit-learn>=1.0.0
torch>=1.10.0
transformers>=4.12.0
twikit>=1.7.6
tqdm>=4.65.0
numpy>=1.26.4
pandas>=2.0.3
praw>=7.8.1
pymongo>=4.6.1
python-dotenv>=1.1.0
werkzeug>=2.0.0